Simplify cleanups
